refactor(snowman): extract test hooks (R3.12)
