+++ docs/ARCHITECTURE.md
@@ -243,10 +243,11 @@ camera.position -= shake                          // revert so smoothing stays c
 > **`src/snowman/model.ts`**, and R3.9 moved heading/tilt/ski-pose animation into
 > **`src/snowman/pose.ts`**. R3.10 moved reset + movement integration into
 > **`src/snowman/physics.ts`**. R3.11 moved tree/rock/bounds/end-of-run checks
-> into **`src/snowman/collision.ts`**. The verification harness self-registers
+> into **`src/snowman/collision.ts`**, and R3.12 moved browser-test hook wiring
+> into **`src/snowman/test-hooks.ts`**. The verification harness self-registers
 > the same `.js` -> `.ts` resolver as the Node suites before importing the facade,
-> so the public seam remains the thing under test as step 12 carves `test-hooks`
-> out of `index.ts`.
+> so the public seam remains the thing under test while `index.ts` keeps the
+> exported snowman contract types and update orchestration.
 
 ---
