+++ src/snowman/index.ts
@@ -20,6 +20,7 @@ import { detectCollisionsAndFinish } from './collision.js';
 import { createSnowman } from './model.js';
 import { resetSnowman, stepSnowmanPhysics } from './physics.js';
 import { applySnowmanPose } from './pose.js';
+import { addTestHooks } from './test-hooks.js';
 
 // These contract types are exported so the typed player-state layer in
 // physics.ts (PR 3.21) shares snowman's exact call contract instead of
@@ -157,164 +158,6 @@ function updateSnowman(snowman: THREE.Object3D, delta: number, pos: PlayerPos, v
 
   // Return updated state variables
   return result;
-}
-
-// Add test hook functions for tree collision testing
-function addTestHooks(pos: PlayerPos, showGameOver: ShowGameOverFn, getTerrainHeight: TerrainHeightFn) {
-  console.log("Snowman.addTestHooks called - setting up test hooks");
-  
-  if (!window.testHooks) {
-    window.testHooks = {};
-  }
-  
-  // Add a force collision function that can be called by tests
-  window.testHooks.forceTreeCollision = function() {
-    console.log("TEST: forceTreeCollision hook called");
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Call the function directly to ensure it works
-    console.log("TEST: Forcing tree collision (direct call)");
-    try {
-      directShowGameOver("BANG!!! You hit a tree!");
-      console.log("TEST: Successfully called showGameOver");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver:", error);
-      // As a fallback, just simulate the collision for the test
-      window.testCollisionDetected = true;
-    }
-    
-    return true;
-  };
-  
-  // Add a tree collision checking function
-  window.testHooks.checkTreeCollision = function(x: number, z: number) {
-    console.log("TEST: checkTreeCollision hook called");
-    // Create a test tree at the specified position
-    const testTree = { 
-      x: x, 
-      y: getTerrainHeight(x, z), 
-      z: z 
-    };
-    
-    // For tree jumping test, check if we're in the air
-    const isJumpingHighAboveTrees = isInAir && verticalVelocity > 0 && pos.y > (testTree.y + 5);
-    console.log(`TEST: Jump check - isInAir=${isInAir}, verticalVelocity=${verticalVelocity}, pos.y=${pos.y}, tree.y=${testTree.y}, jumping=${isJumpingHighAboveTrees}`);
-    
-    // If we're testing tree jumping, we need to handle it properly
-    if (window.testTreeJumpingCheck) {
-      // This is the tree jumping test
-      window.testTreeJumpingCheck = false; // Reset the flag
-      
-      // We should detect the collision even when jumping in the test hook
-      console.log("TEST: checkTreeCollision for jumping test - will detect collision regardless of jumping state");
-      
-      // Create a direct function reference to ensure it works
-      const directShowGameOver = window.showGameOver || showGameOver;
-      
-      // Always detect collision in the jumping test to verify the hook works
-      try {
-        directShowGameOver("BANG!!! You hit a tree (ignoring jump)!");
-        console.log("TEST: Successfully called showGameOver from checkTreeCollision (jumping test)");
-      } catch (error) {
-        console.error("TEST ERROR: Failed to call showGameOver from checkTreeCollision (jumping test):", error);
-        window.testCollisionDetected = true;
-      }
-      return true;
-    }
-    
-    // For regular testing, respect the jumping logic
-    if (isJumpingHighAboveTrees) {
-      console.log("TEST: Snowman is jumping high above trees - no collision");
-      return false;
-    }
-    
-    // Position the snowman directly at the tree for collision testing
-    pos.x = x;
-    pos.z = z;
-    pos.y = testTree.y;
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Check for collision with this test tree
-    console.log(`TEST: Checking collision at (${x}, ${z})`);
-    
-    // Always detect collision in test regardless of distance
-    console.log("TEST: Tree collision detected");
-    try {
-      directShowGameOver("BANG!!! You hit a tree!");
-      console.log("TEST: Successfully called showGameOver from checkTreeCollision");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver from checkTreeCollision:", error);
-      // As a fallback, just simulate the collision for the test
-      window.testCollisionDetected = true;
-    }
-    return true;
-  };
-  
-  // Add a function to test collision in extended terrain area
-  window.testHooks.checkExtendedTerrainCollision = function() {
-    console.log("TEST: checkExtendedTerrainCollision hook called");
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Check if we have any trees in extended terrain
-    if (!window.treePositions || !window.treePositions.length) {
-      console.log("TEST: No trees available for extended terrain test");
-      // Still show collision for test to pass
-      try {
-        directShowGameOver("BANG!!! You hit a tree in extended terrain (simulated)!");
-        console.log("TEST: Successfully called showGameOver for extended terrain");
-      } catch (error) {
-        console.error("TEST ERROR: Failed to call showGameOver for extended terrain:", error);
-        window.testCollisionDetected = true;
-      }
-      return true;
-    }
-    
-    // Find a tree in extended terrain (z < -80)
-    const extendedTrees = window.treePositions.filter((t: { z: number }) => t.z < -80);
-    if (extendedTrees.length === 0) {
-      console.log("TEST: No trees found in extended terrain (z < -80)");
-      // Still show collision for test to pass
-      try {
-        directShowGameOver("BANG!!! You hit a tree in extended terrain (simulated)!");
-        console.log("TEST: Successfully called showGameOver for extended terrain");
-      } catch (error) {
-        console.error("TEST ERROR: Failed to call showGameOver for extended terrain:", error);
-        window.testCollisionDetected = true;
-      }
-      return true;
-    }
-    
-    // Use the first tree in extended terrain for testing
-    const testTree = extendedTrees[0];
-    console.log(`TEST: Using tree at (${testTree.x.toFixed(1)}, ${testTree.z.toFixed(1)}) in extended terrain`);
-    
-    // Position snowman at the tree for collision
-    pos.x = testTree.x;
-    pos.z = testTree.z;
-    pos.y = testTree.y;
-    
-    // Check for collision 
-    console.log("TEST: Positioned snowman directly on extended terrain tree, checking collision");
-    
-    // Always trigger collision in test
-    try {
-      directShowGameOver("BANG!!! You hit a tree in extended terrain!");
-      console.log("TEST: Successfully called showGameOver for extended terrain collision");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver for extended terrain collision:", error);
-      window.testCollisionDetected = true;
-    }
-    return true;
-  };
-  
-  // For debugging in tests
-  console.log("Snowman test hooks installed: forceTreeCollision, checkTreeCollision, checkExtendedTerrainCollision");
 }
 
 // Export snowman functions
